chore: Refactor Comments to CommentForm in Components
